% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_TESEA.R
\name{PlotTESEAPathwayGraph}
\alias{PlotTESEAPathwayGraph}
\title{Plot the pathway-result network diagram}
\usage{
PlotTESEAPathwayGraph(graph, margin = 0, vertex.label.cex = 0.6,
  vertex.label.font = 1, vertex.size = 8, vertex.size2 = 6,
  vertex.shape = "rectangle", layout = igraph::layout.fruchterman.reingold,
  vertex.label.color = "black", edge.color = "dimgray",
  vertex.color = "#C1FFC1", vertex.frame.color = "dimgray", axes = FALSE,
  xlab = "", ylab = "", sub = NULL, main = NULL, ...)
}
\arguments{
\item{graph}{A dataframe of pathway result obtained from the \code{\link{TESEA.Main}} function.}

\item{margin}{A numeric. The value is usually between -0.5 and 0.5, which is able to zoom in or out a pathway graph. The default is 0.}

\item{vertex.label.cex}{A numeric vector of node label size.}

\item{vertex.label.font}{A numeric vector of label font.}

\item{vertex.size}{A numeric vector of Node size. See plot.igraph}

\item{vertex.size2}{A numeric vector of Node size.}

\item{vertex.shape}{A vector of node shape. The default is graphics_type.}

\item{layout}{A matrix of x-y coordinates with two dims. Determine the placement of the nodes for drawing a graph.The default is layout.random.}

\item{vertex.label.color}{A vector of node label colors. The default is black.}

\item{edge.color}{A vector of edge color. The default is dimgray.}

\item{vertex.color}{A vector of node colors. The default is the KEGG node color.}

\item{vertex.frame.color}{A vector of node frame color. The default is dimgray.}

\item{axes}{A logical. whether to plot axes. The default is FALSE.}

\item{xlab}{A character string. The label of the horizontal axis. The default is the empty string.}

\item{ylab}{A character string. The label of the vertical axis. The default is the empty string.}

\item{sub}{A character string of subtitle.}

\item{main}{A character string of main title.}

\item{...}{The arguments passed to or from methods. See plot.igraph and see plot.}
}
\description{
Plot the pathway-result network diagram
}
\author{
Junwei Han, Xinrui Shi and Chunquan Li wrote the original in the ESEA package, small changes by Nello Blaser.
}
