% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_TESEA.R
\name{TESEA.Main}
\alias{TESEA.Main}
\title{Topological edge set enrichment analysis}
\usage{
TESEA.Main(EdgeCorScore, pathwayEdge.db, weighted.score.type = 1,
  pathway = "kegg", gs.size.threshold.min = 15,
  gs.size.threshold.max = 1000, reshuffling.type = "edge.labels",
  nperm = 100, p.val.threshold = -1, FDR.threshold = 0.05, topgs = 1)
}
\arguments{
\item{EdgeCorScore}{A numeric vector. Each element is the differential correlation score of an edge.}

\item{pathwayEdge.db}{A character vector, the length of it is the number of pathways.}

\item{weighted.score.type}{A value. Edge enrichment correlation-based weighting: 0=no weight, 1=standard weigth, 2 = over-weigth. The default value is 1}

\item{pathway}{A character string of pathway database. Should be one of "kegg","reactome", "nci","huamncyc","biocarta","spike" and "panther". The default value is "kegg"}

\item{gs.size.threshold.min}{An integer. The minimum size (in edges) for pathways to be considered. The default value is 15.}

\item{gs.size.threshold.max}{An integer. The maximum size (in edges) for pathways to be considered. The default value is 1000.}

\item{reshuffling.type}{A character string. The type of permutation reshuffling: "edge.labels" or "gene.labels". The default value is "edge.labels".}

\item{nperm}{An integer. The number of permutation reshuffling. The default value is 100.}

\item{p.val.threshold}{A value. The significance threshold of NOM p-value for pathways whose detail results of pathways to be presented. The  default value is -1, which means no threshold.}

\item{FDR.threshold}{A value. The significance threshold of FDR q-value for pathways whose detail results of pathways to be presented. The default value is 0.05.}

\item{topgs}{An integer. The number of top scoring gene sets used for detailed reports. The default value is 1.}
}
\description{
Topological edge set enrichment analysis
}
\author{
Junwei Han, Xinrui Shi and Chunquan Li wrote the original in the ESEA package, small changes by Nello Blaser.
}
