% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_score.R
\name{triangle_creation_score}
\alias{triangle_creation_score}
\title{Triangle creation score}
\usage{
triangle_creation_score(dataset, class.labels, controlcharacter,
  edgesbackground, return_p = TRUE, metric = "angular")
}
\arguments{
\item{dataset}{A marix of gene expression data whose row names are genes 
symbols and whose column names are samples.}

\item{class.labels}{A vector of binary labels.}

\item{controlcharacter}{A character string of control sample label.}

\item{edgesbackground}{A marix which deposits the data of background set of 
edges.}

\item{return_p}{Should p value attributes be calculated.}

\item{metric}{The distance metric to use. Defaults to "angular".}
}
\description{
Calculate the difference between the proportions of triangles an edge 
creates in the exposed group versus the control group. This is an 
alternative to \code{\link[ESEA]{calEdgeCorScore}} from the \code{ESEA} package 
to use in \code{\link{TESEA.Main}}.
}
