% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_TESEA.R
\name{TESEA.Main}
\alias{TESEA.Main}
\title{Topological edge set enrichment analysis}
\usage{
TESEA.Main(EdgeScore, pathwayEdge.db, weighted.score.type = 1,
  pathway = "kegg", gs.size.threshold.min = 15,
  gs.size.threshold.max = 1000, reshuffling.type = "edge.labels",
  nperm = 100, p.val.threshold = -1, FDR.threshold = 0.05, topgs = 1)
}
\arguments{
\item{EdgeScore}{A numeric vector. Each element is the differential score of an edge from \code{\link{triangle_creation_score}} or \code{\link[ESEA]{calEdgeCorScore}}.}

\item{pathwayEdge.db}{A character vector, the length of it is the number of pathways.}

\item{weighted.score.type}{A value. Edge enrichment correlation-based weighting: 0=no weight, 1=standard weigth, 2 = over-weigth. The default value is 1}

\item{pathway}{A character string of pathway database. Should be one of "kegg","reactome", "nci","huamncyc","biocarta","spike" and "panther". The default value is "kegg"}

\item{gs.size.threshold.min}{An integer. The minimum size (in edges) for pathways to be considered. The default value is 15.}

\item{gs.size.threshold.max}{An integer. The maximum size (in edges) for pathways to be considered. The default value is 1000.}

\item{reshuffling.type}{A character string. The type of permutation reshuffling: "edge.labels" or "gene.labels". The default value is "edge.labels".}

\item{nperm}{An integer. The number of permutation reshuffling. The default value is 100.}

\item{p.val.threshold}{A value. The significance threshold of NOM p-value for pathways whose detail results of pathways to be presented. The  default value is -1, which means no threshold.}

\item{FDR.threshold}{A value. The significance threshold of FDR q-value for pathways whose detail results of pathways to be presented. The default value is 0.05.}

\item{topgs}{An integer. The number of top scoring gene sets used for detailed reports. The default value is 1.}
}
\description{
Topological edge set enrichment analysis
}
\examples{
\dontrun{
# set random seed
set.seed(1)
# get data from ESEA
ESEA::initializeESEA()
edgesbackground <- ESEA::GetEdgesBackgrandData()
pathwayEdge.db <- ESEA::GetPathwayEdgeData()
dataset <- ESEA::GetExampleData("dataset")
class.labels <- ESEA::GetExampleData("class.labels")
controlcharacter <- ESEA::GetExampleData("controlcharactor") 
# calculate edge score (triangle version)
EdgeTriScore <- triangle_creation_score(dataset,
                                        class.labels, 
                                        controlcharacter, 
                                        edgesbackground)
# topological edge set enrichment analysis
Results_Tri <- TESEA.Main(
  EdgeTriScore, 
  pathwayEdge.db, weighted.score.type = 1,
  pathway = "kegg", gs.size.threshold.min = 15,
  gs.size.threshold.max = 1000, 
  reshuffling.type = "edge.labels", nperm = 1000,
  p.val.threshold= -1, FDR.threshold = 0.05, topgs =1)
# look at results
res_Tri <- Results_Tri[["summary"]]
res_Tri[res_Tri[, 'FDR q-val'] < 0.05, c("GS", "SIZE", "ES", "NES")]
##                                       GS SIZE       ES       NES
## 103 Maturity onset diabetes of the young   17  0.90011  2.373214
## 15   Antigen processing and presentation   56 -0.54973 -1.931995
## 126                p53 signaling pathway   70  0.46139  1.638150
## 147               Rap1 signaling pathway  980 -0.27030 -1.296834
## 148                Ras signaling pathway  976 -0.26085 -1.245997
# plot first pathway
PlotTESEAPathwayGraph(Results_Tri[['pathways']][[1]])
}
}
\author{
Junwei Han, Xinrui Shi and Chunquan Li wrote the original in the ESEA package, small changes by Nello Blaser.
}
